requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
anyascii==0.3.0
    # via wagtail
asgiref==3.5.0
    # via django
beautifulsoup4==4.9.3
    # via wagtail
boto3==1.20.26
    # via -r requirements.in
botocore==1.23.46
    # via
    #   boto3
    #   s3transfer
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.10
    # via requests
cryptography==36.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==0.5.0
    # via -r requirements.in
django-allauth==0.47.0
    # via -r requirements.in
django-anymail[mailgun]==8.4
    # via -r requirements.in
django-filter==21.1
    # via wagtail
django-modelcluster==5.2
    # via wagtail
django-storage-url==0.8.1
    # via -r requirements.in
django-storages==1.12.3
    # via django-storage-url
django-taggit==1.5.1
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
django==3.2.10
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-anymail
    #   django-filter
    #   django-storage-url
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
djangorestframework==3.13.1
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
flickrapi==2.4.0
    # via -r requirements.in
furl==2.1.3
    # via django-storage-url
html5lib==1.1
    # via wagtail
idna==3.3
    # via requests
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
l18n==2021.3
    # via wagtail
oauthlib==3.2.0
    # via requests-oauthlib
openpyxl==3.0.9
    # via tablib
orderedmultidict==1.0.1
    # via furl
pillow==8.4.0
    # via
    #   -r requirements.in
    #   wagtail
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
requests-oauthlib==1.3.1
    # via
    #   django-allauth
    #   flickrapi
requests-toolbelt==0.9.1
    # via flickrapi
requests==2.27.1
    # via
    #   django-allauth
    #   django-anymail
    #   flickrapi
    #   requests-oauthlib
    #   requests-toolbelt
    #   wagtail
s3transfer==0.5.0
    # via boto3
sentry-sdk==1.5.1
    # via -r requirements.in
six==1.16.0
    # via
    #   flickrapi
    #   furl
    #   html5lib
    #   l18n
    #   orderedmultidict
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tablib[xls,xlsx]==3.2.0
    # via wagtail
telepath==0.2
    # via wagtail
urllib3==1.26.8
    # via
    #   botocore
    #   requests
    #   sentry-sdk
uwsgi==2.0.20
    # via -r requirements.in
wagtail==2.15.1
    # via -r requirements.in
webencodings==0.5.1
    # via html5lib
whitenoise==5.3.0
    # via -r requirements.in
willow==1.4
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.2
    # via wagtail
xlwt==1.3.0
    # via tablib
